feat: turned `title` field to unique, improved repository URL validation and started tests implementation (WIP)

+++ challenge-management-api/src/challenge/challenge.service.ts
@@ -1,9 +1,7 @@
-import { Inject, Injectable } from '@nestjs/common';
+import { BadRequestException, Inject, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateChallengeInput } from './dto/create-challenge.input';
 import { UpdateChallengeInput } from './dto/update-challenge.input';
-import { PrismaService } from '../prisma/prisma.service';
 import { ListChallengesArgs } from './dto/list-challenges.args';
-import { parseFilters } from '../utils/parse-filters';
 import { ChallengeRepository } from './repositories/challenge.repository';
 
 @Injectable()
@@ -12,11 +10,15 @@ export class ChallengeService {
     @Inject('ChallengeRepository')
     private readonly challengeRepository: ChallengeRepository
   ) {}
-  create(createChallengeInput: CreateChallengeInput) {
+  async create(createChallengeInput: CreateChallengeInput) {
+    const existingChallenge = await this.challengeRepository.findByTitle(createChallengeInput.title);
+    if (existingChallenge) {
+      throw new BadRequestException(`challenge with title ${createChallengeInput.title} already exists!`);
+    }
     return this.challengeRepository.create(createChallengeInput);
   }
 
-  findAll(args: ListChallengesArgs) {
+  async findAll(args: ListChallengesArgs) {
     let { page, limit, title, description } = args;
     const defaultPageSize = 15;
     limit = limit ?? defaultPageSize;
@@ -27,19 +29,26 @@ export class ChallengeService {
       title,
       description,
     };
-
     return this.challengeRepository.findMany(queryArgs);
   }
 
-  findOne(id: string) {
+  async findOne(id: string) {
     return this.challengeRepository.findOne(id);
   }
 
-  update(updateChallengeInput: UpdateChallengeInput) {
+  async update(updateChallengeInput: UpdateChallengeInput) {
+    const existingChallenge = await this.challengeRepository.findByTitle(updateChallengeInput.title);
+    if (existingChallenge) {
+      throw new BadRequestException(`challenge with title ${updateChallengeInput.title} already exists!`);
+    }
     return this.challengeRepository.update(updateChallengeInput);
   }
 
-  remove(id: string) {
+  async delete(id: string) {
+    const challengeExists = await this.challengeRepository.findOne(id);
+    if (!challengeExists) {
+      throw new NotFoundException();
+    }
     return this.challengeRepository.delete(id);
   }
 }
